refactor(verifier-v2): clean + consolidate interface/fcts into one file
