+++ encoding/kzg/verifier/v2/verifier_test.go
@@ -11,8 +11,136 @@ import (
 	"github.com/Layr-Labs/eigenda/encoding/kzg/committer"
 	"github.com/Layr-Labs/eigenda/encoding/kzg/prover/v2"
 	"github.com/Layr-Labs/eigenda/encoding/kzg/verifier/v2"
+	"github.com/Layr-Labs/eigenda/encoding/rs"
+
 	"github.com/stretchr/testify/require"
 )
+
+func TestVerifyFrames(t *testing.T) {
+	harness := getTestHarness()
+
+	params := encoding.ParamsFromSysPar(harness.numSys, harness.numPar, uint64(len(harness.paddedGettysburgAddressBytes)))
+
+	proverGroup, err := prover.NewProver(harness.proverV2KzgConfig, nil)
+	require.Nil(t, err)
+
+	committer, err := committer.NewFromConfig(*harness.committerConfig)
+	require.Nil(t, err)
+
+	frames, err := proverGroup.GetFrames(harness.paddedGettysburgAddressBytes, params)
+	require.Nil(t, err)
+	commitments, err := committer.GetCommitmentsForPaddedLength(harness.paddedGettysburgAddressBytes)
+	require.Nil(t, err)
+
+	verifierGroup, err := verifier.NewVerifier(harness.verifierV2KzgConfig, nil)
+	require.Nil(t, err)
+
+	indices := []encoding.ChunkNumber{}
+	for i := range len(frames) {
+		indices = append(indices, encoding.ChunkNumber(i))
+	}
+	err = verifierGroup.VerifyFrames(frames, indices, commitments, params)
+	require.Nil(t, err)
+}
+
+func TestUniversalVerify(t *testing.T) {
+	harness := getTestHarness()
+
+	group, err := prover.NewProver(harness.proverV2KzgConfig, nil)
+	require.Nil(t, err)
+
+	committer, err := committer.NewFromConfig(*harness.committerConfig)
+	require.Nil(t, err)
+
+	v, err := verifier.NewVerifier(harness.verifierV2KzgConfig, nil)
+	require.Nil(t, err)
+
+	params := encoding.ParamsFromSysPar(harness.numSys, harness.numPar, uint64(len(harness.paddedGettysburgAddressBytes)))
+	enc, err := group.GetKzgEncoder(params)
+	require.Nil(t, err)
+
+	numBlob := 5
+	samples := make([]encoding.Sample, 0)
+	for z := 0; z < numBlob; z++ {
+		inputFr, err := rs.ToFrArray(harness.paddedGettysburgAddressBytes)
+		require.Nil(t, err)
+
+		commit, _, _, err := committer.GetCommitments(inputFr)
+		require.Nil(t, err)
+		frames, fIndices, err := enc.GetFrames(inputFr)
+		require.Nil(t, err)
+
+		// create samples
+		for i := 0; i < len(frames); i++ {
+			f := frames[i]
+			j := fIndices[i]
+
+			q, err := rs.GetLeadingCosetIndex(uint64(i), harness.numSys+harness.numPar)
+			require.Nil(t, err)
+
+			require.Equal(t, j, q, "leading coset inconsistency")
+
+			sample := encoding.Sample{
+				Commitment:      (*encoding.G1Commitment)(commit),
+				Chunk:           &f,
+				BlobIndex:       z,
+				AssignmentIndex: uint(i),
+			}
+			samples = append(samples, sample)
+		}
+	}
+
+	require.True(t, v.UniversalVerifySubBatch(params, samples, numBlob) == nil, "universal batch verification failed\n")
+}
+
+func TestUniversalVerifyWithPowerOf2G2(t *testing.T) {
+	harness := getTestHarness()
+	group, err := prover.NewProver(harness.proverV2KzgConfig, nil)
+	require.Nil(t, err)
+
+	committer, err := committer.NewFromConfig(*harness.committerConfig)
+	require.Nil(t, err)
+
+	v, err := verifier.NewVerifier(harness.verifierV2KzgConfig, nil)
+	require.NoError(t, err)
+
+	params := encoding.ParamsFromSysPar(harness.numSys, harness.numPar, uint64(len(harness.paddedGettysburgAddressBytes)))
+	enc, err := group.GetKzgEncoder(params)
+	require.NoError(t, err)
+
+	numBlob := 5
+	samples := make([]encoding.Sample, 0)
+	for z := 0; z < numBlob; z++ {
+		inputFr, err := rs.ToFrArray(harness.paddedGettysburgAddressBytes)
+		require.Nil(t, err)
+
+		commit, _, _, err := committer.GetCommitments(inputFr)
+		require.Nil(t, err)
+		frames, fIndices, err := enc.GetFrames(inputFr)
+		require.Nil(t, err)
+
+		// create samples
+		for i := 0; i < len(frames); i++ {
+			f := frames[i]
+			j := fIndices[i]
+
+			q, err := rs.GetLeadingCosetIndex(uint64(i), harness.numSys+harness.numPar)
+			require.Nil(t, err)
+
+			require.Equal(t, j, q, "leading coset inconsistency")
+
+			sample := encoding.Sample{
+				Commitment:      (*encoding.G1Commitment)(commit),
+				Chunk:           &f,
+				BlobIndex:       z,
+				AssignmentIndex: uint(i),
+			}
+			samples = append(samples, sample)
+		}
+	}
+
+	require.True(t, v.UniversalVerifySubBatch(params, samples, numBlob) == nil, "universal batch verification failed\n")
+}
 
 func TestBenchmarkVerifyChunks(t *testing.T) {
 	t.Skip("This test is meant to be run manually, not as part of the test suite")
